Implement MapProperty and fix bugs
Bug that was fixed:

An array of structs can have None values (I think), so while reading an NestedValue,
if the Property for a Complex is None, then reset the Cursor position and add
a None to the array.

This may break for missing values on properties that are not the last property
in a file. Or it might just be wrong for Arrays of Structs.