Refactor properties loading
A new concept, a Loader, is introduced that holds the logic for deserializaing
and serializaing a property. Loaders are implemented in asset::property::loaders
and are uses to parse properties.

This new architecture makes it much easier to split up code into smaller more
understandable pieces, and makes adding new property types easier.